requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
openai>=1.12.0
python-dotenv==1.0.0
werkzeug==3.0.1
gunicorn==21.2.0
cloud-sql-python-connector[pymysql]==1.11.0
pymysql==1.1.0

